backend/requirements.txt
# ══════════════════════════════════════════════════════════════════════════════
# SkillForge — Backend Dependencies
# Final merged requirements.txt
# Python 3.11+ required
# Install: pip install -r requirements.txt
# ══════════════════════════════════════════════════════════════════════════════

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ── PDF Parsing ───────────────────────────────────────────────────────────────
pymupdf==1.23.26

# ── AI / ML ───────────────────────────────────────────────────────────────────
spacy
sentence-transformers==2.7.0
faiss-cpu==1.13.2
networkx==3.3
scikit-learn==1.4.2
numpy==1.26.4

# ── LLM ───────────────────────────────────────────────────────────────────────
groq==0.9.0

# ── HTTP ──────────────────────────────────────────────────────────────────────
requests==2.32.3
httpx==0.27.0

# ── Config ────────────────────────────────────────────────────────────────────
pydantic-settings==2.2.1
python-dotenv==1.0.1